requirements.txt
Flask==3.0.3
flask-login
Flask-SQLAlchemy
psycopg2-binary
python-dotenv
Werkzeug